refactor: finish cmake restructure
