--- .vscode/settings.json
@@ -1,83 +1,0 @@
-{
-    "files.associations": {
-        "iostream": "cpp",
-        "limits": "cpp",
-        "vector": "cpp",
-        "__split_buffer": "cpp",
-        "string": "cpp",
-        "ios": "cpp",
-        "filesystem": "cpp",
-        "algorithm": "cpp",
-        "__bit_reference": "cpp",
-        "__bits": "cpp",
-        "__config": "cpp",
-        "__debug": "cpp",
-        "__errc": "cpp",
-        "__locale": "cpp",
-        "__mutex_base": "cpp",
-        "__nullptr": "cpp",
-        "__string": "cpp",
-        "__threading_support": "cpp",
-        "__tuple": "cpp",
-        "array": "cpp",
-        "atomic": "cpp",
-        "bitset": "cpp",
-        "cctype": "cpp",
-        "chrono": "cpp",
-        "clocale": "cpp",
-        "cmath": "cpp",
-        "compare": "cpp",
-        "complex": "cpp",
-        "concepts": "cpp",
-        "cstdarg": "cpp",
-        "cstddef": "cpp",
-        "cstdint": "cpp",
-        "cstdio": "cpp",
-        "cstdlib": "cpp",
-        "cstring": "cpp",
-        "ctime": "cpp",
-        "cwchar": "cpp",
-        "cwctype": "cpp",
-        "exception": "cpp",
-        "initializer_list": "cpp",
-        "iosfwd": "cpp",
-        "istream": "cpp",
-        "locale": "cpp",
-        "memory": "cpp",
-        "mutex": "cpp",
-        "new": "cpp",
-        "ostream": "cpp",
-        "ratio": "cpp",
-        "sstream": "cpp",
-        "stdexcept": "cpp",
-        "streambuf": "cpp",
-        "string_view": "cpp",
-        "system_error": "cpp",
-        "tuple": "cpp",
-        "type_traits": "cpp",
-        "typeinfo": "cpp",
-        "variant": "cpp",
-        "iterator": "cpp",
-        "utility": "cpp",
-        "__hash_table": "cpp",
-        "__node_handle": "cpp",
-        "__tree": "cpp",
-        "csignal": "cpp",
-        "deque": "cpp",
-        "fstream": "cpp",
-        "iomanip": "cpp",
-        "map": "cpp",
-        "set": "cpp",
-        "stack": "cpp",
-        "unordered_set": "cpp",
-        "any": "cpp",
-        "forward_list": "cpp",
-        "numeric": "cpp",
-        "ranges": "cpp",
-        "unordered_map": "cpp",
-        "valarray": "cpp",
-        "optional": "cpp",
-        "list": "cpp",
-        "typeindex": "cpp"
-    }
-}
