+++ libseam_carving/CMakeLists.txt
@@ -1,0 +1,18 @@
+set(
+    LIB_SRCS
+        energy.cpp coord.cpp seam.cpp carver.cpp)
+
+add_library(seam_carving SHARED ${LIB_SRCS})
+
+find_package(OpenCV CONFIG REQUIRED)
+
+target_include_directories(
+        seam_carving
+        PUBLIC seam_carving
+        PRIVATE seam_carving/tests
+)
+
+target_link_libraries(
+    seam_carving
+        PRIVATE ${OPENCV_LIBS}
+)
