+++ libseam_carving/seam_carving/tests/carver_data.hpp
@@ -1,20 +1,30 @@
-#include "Seam.h"
-#include "TestData.h"
+//
+// Created by Andy Mina on 3/7/23.
+//
 
+#ifndef SEAM_CARVING_CARVER_DATA_HPP
+#define SEAM_CARVING_CARVER_DATA_HPP
+
+// project
+#include <seam_carving/seam.hpp>
+#include <seam_carving/tests/test_data.hpp>
+
+// 3rd party
+#include <nlohmann/json.hpp>
 #include <opencv2/core.hpp>
 
-namespace SeamCarving {
+namespace seam_carving {
     struct CarverTestData : TestData {
-        Seam vertical_seam, horiztonal_seam;        
+        Seam vertical_seam, horizontal_seam;
         cv::Mat remove_vertical_matrix, remove_horizontal_matrix;
-        cv::Mat insert_vertical_matrix, insert_horizontal_matrix;        
+        cv::Mat insert_vertical_matrix, insert_horizontal_matrix;
     };
 
     inline void from_json(const nlohmann::json& j, CarverTestData& carver_data) {
         const nlohmann::json& root = j.at("carver");
-        
+
         carver_data.vertical_seam = root.at("find - vertical");
-        carver_data.horiztonal_seam =  root.at("find - horizontal");
+        carver_data.horizontal_seam =  root.at("find - horizontal");
 
         carver_data.remove_vertical_matrix = root.at("remove - vertical");
         carver_data.remove_horizontal_matrix = root.at("remove - horizontal");
@@ -22,3 +32,5 @@ namespace SeamCarving {
         carver_data.insert_horizontal_matrix = root.at("insert - horizontal");
     }
 }
+
+#endif //SEAM_CARVING_CARVER_DATA_HPP
