+++ libseam_carving/tests/coord.test.cpp
@@ -1,0 +1,3 @@
+//
+// Created by Andy Mina on 3/7/23.
+//
