+++ libseam_carving/tests/data/Seam Carving Tests - TestId_0.tsv
