+++ libseam_carving/tests/data/Seam Carving Tests - TestId_1.tsv
