+++ libseam_carving/tests/data/Seam Carving Tests - TestId_2.tsv
