+++ libseam_carving/tests/data/Seam Carving Tests - TestId_3.tsv
