+++ libseam_carving/tests/data/Seam Carving Tests - TestId_4.tsv
