+++ seam_carving/main.cpp
@@ -1,0 +1,12 @@
+//
+// Created by Andy Mina on 3/7/23.
+//
+
+#include <iostream>
+#include <seam_carving/energy.hpp>
+#include <opencv2/core.hpp>
+
+int main() {
+    std::cout << "hello world\n";
+    return 0;
+}
