+++ vcpkg.json
@@ -1,0 +1,12 @@
+{
+  "name" : "seam-carving",
+  "version-string" : "1.0.0",
+  "builtin-baseline" : "e69c9817d28e0ad37a79b5623c17c2b87df8939f",
+  "dependencies" : [ {
+    "name" : "nlohmann-json",
+    "version>=" : "3.11.2"
+  }, {
+    "name" : "opencv",
+    "version>=" : "4.7.0#1"
+  } ]
+}
